{{cookiecutter.project_name}}/{{cookiecutter.app_name}}/models/user.py
from sqlalchemy.ext.hybrid import hybrid_property

from {{cookiecutter.app_name}}.extensions import db, pwd_context
from {{cookiecutter.app_name}}.extensions import bcrypt
import datetime
from flask import current_app


class User(db.Model):
    """Basic user model"""

    # __tablename__ = "users"

    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
    username = db.Column(db.String(80), unique=True, nullable=False)
    email = db.Column(db.String(80), unique=True, nullable=False)
    _password = db.Column("password", db.String(255), nullable=False)
    registered_on = db.Column(db.DateTime, nullable=False)
    admin = db.Column(db.Boolean, nullable=False, default=False)
    active = db.Column(db.Boolean, default=True)

    def __init__(self, username, email, password, active, admin=False):
        self.username = username
        self.email = email
        self.password = bcrypt.generate_password_hash(
            password, current_app.config.get("BCRYPT_LOG_ROUNDS")
        ).decode("utf-8")
        self.registered_on = datetime.datetime.now()
        self.active = active
        self.admin = admin
    
    @hybrid_property
    def password(self):
        return self._password

    @password.setter
    def password(self, value):
        self._password = pwd_context.hash(value)

    def __repr__(self):
        return "<User %s>" % self.username
